teset case passed

+++ springapp/src/main/java/com/examly/springapp/repository/ProfileRepo.java
@@ -6,5 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProfileRepo extends JpaRepository<Profile,Integer> {
+public interface ProfileRepo extends JpaRepository<Profile,Long> {
+    
 }
